requirements.txt
Django==1.11.7
django-address==0.1.6
django-admin-sortable2==0.6.16
django-crispy-forms==1.7.0
django-mysql==2.1.1
django-sortedm2m==1.5.0
django-sortedm2m-filter-horizontal-widget==1.2
mysqlclient==1.3.12
pytz==2017.3
